+++ client/src/app/app.module.ts
@@ -19,7 +19,7 @@ import { DeviceComponent } from "./views/devices/device/device.component";
 import { DatasetComponent } from "./views/datasets/dataset/dataset.component";
 import { DetailsComponent } from "./views/datasets/dataset/details/details.component";
 import { MatTabsModule } from "@angular/material/tabs";
-import { DataComponent } from "./views/datasets/dataset/data/data.component";
+import { LineChartNgxComponent } from "./views/datasets/dataset/line-chart-ngx/line-chart-ngx.component";
 import { NewDeviceDialogComponent } from "./dialogs/new-device/new-device.component";
 import { MatDialogModule } from "@angular/material/dialog";
 import { MatCardModule } from "@angular/material/card";
@@ -48,7 +48,7 @@ import { FlexLayoutModule } from "@angular/flex-layout";
     DeviceComponent,
     DatasetComponent,
     DetailsComponent,
-    DataComponent,
+    LineChartNgxComponent,
     NewDeviceDialogComponent,
     NewDatasetDialogComponent,
     LineChartComponent,
