+++ client/src/app/views/datasets/dataset/line-chart-ngx/line-chart-ngx.component.ts
@@ -1,14 +1,14 @@
 import { Component, Input, OnInit } from "@angular/core";
 import { Dataset, Sensordata } from "src/app/models/dataset";
 import { DatasetsService } from "src/app/services/datasets.service";
-import { multi, SAMPLEDATA } from "./data";
+import { multi, SAMPLEDATA } from "./line-chart-ngx";
 
 @Component({
   selector: "app-dataset-data",
-  templateUrl: "./data.component.html",
-  styleUrls: ["./data.component.scss"],
+  templateUrl: "./line-chart-ngx.component.html",
+  styleUrls: ["./line-chart-ngx.component.scss"],
 })
-export class DataComponent implements OnInit {
+export class LineChartNgxComponent implements OnInit {
   @Input() dataset: Dataset;
   data: any;
   loading: boolean = true;
